fix: :truck: Moved Wall spawn to src/spawn/wall.rs
Issue: #39
